moved the functionalety to useTypingPractice hook for typing practice functionality
